Refactor extension to work in 4.2
