--- nb_anacondacloud/nbextension/__init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from .handlers import load_jupyter_server_extension
